+++ frontend/src/lib/components/apps/editor/component/Component.svelte
@@ -166,10 +166,11 @@
 	}}
 	on:mouseout|stopPropagation={mouseOut}
 	class={twMerge(
-		'h-full flex flex-col w-full component relative',
+		'h-full flex flex-col w-full component relative connection-access',
 		initializing ? 'overflow-hidden h-0' : '',
 		hidden && $mode === 'preview' ? 'hidden' : ''
 	)}
+	data-connection-button
 >
 	{#if locked && componentActive && $componentActive && moveMode === 'move' && componentDraggedId && componentDraggedId !== component.id && cachedAreOnTheSameSubgrid}
 		<div
@@ -237,10 +238,10 @@
 			$mode === 'dnd' ? 'bg-surface/40' : '',
 			$hoverStore === component.id && $mode !== 'preview'
 				? $connectingInput.opened
-					? 'outline outline-orange-600'
-					: 'outline outline-blue-600'
+					? 'outline outline-[#f8aa4b]'
+					: 'outline outline-blue-400'
 				: '',
-			selected && $mode !== 'preview' ? 'outline outline-indigo-600' : '',
+			selected && $mode !== 'preview' ? 'outline outline-blue-600' : '',
 			$mode != 'preview' ? 'cursor-pointer' : '',
 			'relative z-auto',
 			$app.css?.['app']?.['component']?.class,