+++ frontend/src/lib/components/apps/editor/contextPanel/ComponentOutputViewer.svelte
@@ -41,12 +41,14 @@
 
 {#if object != undefined && Object.keys(object).length > 0}
 	{#if $hasResult[componentId] || $search == ''}
-		<div class="pl-2">
+		<div class="pl-2 !cursor-pointer connection-access" data-connection-button>
 			<ObjectViewer
 				json={filtered}
 				on:select
 				topBrackets={false}
 				pureViewer={!$connectingInput.opened}
+				allowCopy={!$connectingInput.opened}
+				connecting={$connectingInput.opened}
 			/>
 		</div>
 	{:else if $search.length > 0}