+++ frontend/src/lib/components/details/DetailPageDetailPanel.svelte
@@ -51,6 +51,7 @@
 						<slot slot="webhooks" name="webhooks" />
 						<slot slot="routes" name="routes" />
 						<slot slot="websockets" name="websockets" />
+						<slot slot="kafka" name="kafka" />
 						<slot slot="emails" name="emails" />
 						<slot slot="schedules" name="schedules" />
 						<slot slot="cli" name="cli" />