+++ frontend/src/lib/components/graph/renderers/triggers/TriggersBadge.svelte
@@ -9,7 +9,8 @@
 	import { getContext } from 'svelte'
 	import { type TriggerContext } from '$lib/components/triggers'
 	import { FlowService, ScriptService } from '$lib/gen'
-	import { workspaceStore } from '$lib/stores'
+	import { enterpriseLicense, workspaceStore } from '$lib/stores'
+	import KafkaIcon from '$lib/components/icons/KafkaIcon.svelte'
 
 	const { selectedTrigger, triggersCount } = getContext<TriggerContext>('TriggerContext')
 
@@ -18,8 +19,14 @@
 	export let isFlow: boolean
 	export let selected: boolean
 	export let showOnlyWithCount: boolean
-	export let triggersToDisplay: ('webhooks' | 'schedules' | 'routes' | 'websockets' | 'emails')[] =
-		['webhooks', 'schedules', 'routes', 'websockets', 'emails']
+	export let triggersToDisplay: (
+		| 'webhooks'
+		| 'schedules'
+		| 'routes'
+		| 'websockets'
+		| 'kafka'
+		| 'emails'
+	)[] = ['webhooks', 'schedules', 'routes', 'websockets', 'kafka', 'emails']
 	const dispatch = createEventDispatcher()
 
 	onMount(() => {
@@ -47,13 +54,14 @@
 		schedules: { icon: Calendar, countKey: 'schedule_count' },
 		routes: { icon: Route, countKey: 'http_routes_count' },
 		websockets: { icon: Unplug, countKey: 'websocket_count' },
+		kafka: { icon: KafkaIcon, countKey: 'kafka_count' },
 		emails: { icon: Mail, countKey: 'email_count' }
 	}
 </script>
 
 {#each triggersToDisplay as type}
 	{@const { icon, countKey } = triggerTypeConfig[type]}
-	{#if !showOnlyWithCount || ($triggersCount?.[countKey] ?? 0) > 0}
+	{#if (!showOnlyWithCount || ($triggersCount?.[countKey] ?? 0) > 0) && !(type === 'kafka' && !$enterpriseLicense)}
 		<Popover>
 			<svelte:fragment slot="text">{type.charAt(0).toUpperCase() + type.slice(1)}</svelte:fragment>
 			<TriggerButton