+++ frontend/src/lib/components/triggers.ts
@@ -11,7 +11,14 @@ export type ScheduleTrigger = {
 
 export type TriggerContext = {
 	selectedTrigger: Writable<
-		'webhooks' | 'emails' | 'schedules' | 'cli' | 'routes' | 'websockets' | 'scheduledPoll'
+		| 'webhooks'
+		| 'emails'
+		| 'schedules'
+		| 'cli'
+		| 'routes'
+		| 'websockets'
+		| 'scheduledPoll'
+		| 'kafka'
 	>
 	primarySchedule: Writable<ScheduleTrigger | undefined | false>
 	triggersCount: Writable<TriggersCount | undefined>